requirements.txt
asgiref==3.5.0
Babel==2.10.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
defusedxml==0.7.1
Django==4.0.4
django-filter==21.1
django-money==2.1.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filters==1.3.2
filters-django==1.0.5
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
py-moneyed==1.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
